Add babel transpilation
Move source code from lib to src. lib will be git ignored in next commit.

Replace root index.js with specified main in package.json
