+++ README.md
@@ -6,6 +6,7 @@
 * [x] Replace jscs & jshint with eslint & prettier
 * [x] Upgrade dependencies
 * [ ] Update to modern class based JS with Babel
-  * [ ] Add babel transpilation
+  * [x] Add babel transpilation
+  * [ ] Update to modern syntax
   * [ ] Replace bluebird with native promises
 * [ ] Upgrade to TS
