+++ src/cursor.js
