+++ src/inmemory/collection.js
