+++ src/merger.js
